fix: show slide intro in initial screen
